Fix usage of mapserver function that changed result
Removed workaround for filter in ows that rewrote query; Doesn't work
anymore, because getFilterString of mapserver now returns some string
that is not SQL and was missing the original filter (still written in
SQL).
Also removed layerAuthorization check because it used the same method
and did not properly work to begin with.
Added scream test Exception to potentially insecure functions